refactor: move stream creation functions to utils folder
